separate config parsing and add periodic callback to sync databases
