+++ forest/cli/main.py
@@ -67,5 +67,3 @@ def add_bokeh_arguments(parser):
     parser.add_argument(
         "--allow-websocket-origin", metavar="HOST[:PORT]",
         help="public hostnames that may connect to the websocket")
-
-
