+++ forest/main.py
@@ -48,11 +48,9 @@ def map_figure(x_range, y_range):
     return figure
 
 
-def main(argv=None):
-
+def configure(argv=None):
     args = parse_args.parse_args(argv)
     data.AUTO_SHUTDOWN = args.auto_shutdown
-    
     if len(args.files) > 0:
         if args.config_file is not None:
             raise Exception('--config-file and [FILE [FILE ...]] not compatible')
@@ -63,6 +61,11 @@ def main(argv=None):
                 variables=cfg.combine_variables(
                     os.environ,
                     args.variables))
+    return config
+
+
+def main(argv=None):
+    config = configure(argv=argv)
 
     # Feature toggles
     if "feature" in config.plugins:
